Newtheme (#18)
* Prettify documentation & styling updates

Code of conduct and related pages now have styling and navigation! _site/ has been added to .gitignore! The checkbox for the signup form is in a better UI position! The headline is more prominent!

* Oy. not done. needs more work!

* New theme ready to go + RFC docs

There are some redundant files and minor media query styling improvments that can be done. Issus

* Fixed url and favicon

* Added twitter link and email address

Fixes: #5